+++ packages/main/src/RPA/Email/Exchange.py
@@ -1,38 +1,35 @@
 import datetime
+import email
 import logging
-from multiprocessing import AuthenticationError
-import os
-from pathlib import Path
 import re
 import time
-from typing import Any, List, Optional, Union
-import email
 from email import policy  # pylint: disable=no-name-in-module
-import pytz
+from multiprocessing import AuthenticationError
+from pathlib import Path
+from typing import Any, Callable, Dict, List, Optional, Union
 
+import pytz
 from exchangelib import (
+    DELEGATE,
+    IMPERSONATION,
+    OAUTH2,
+    UTC,
     Account,
     Configuration,
     Credentials,
-    DELEGATE,
     FileAttachment,
     Folder,
     HTMLBody,
-    IMPERSONATION,
+    Identity,
     ItemAttachment,
     Mailbox,
     Message,
-    UTC,
+    OAuth2AuthorizationCodeCredentials,
 )
+from exchangelib.folders import Inbox
+from oauthlib.oauth2 import OAuth2Token
 
-
-def mailbox_to_email_address(mailbox):
-    return {
-        "name": mailbox.name if hasattr(mailbox, "name") else "",
-        "email_address": mailbox.email_address
-        if hasattr(mailbox, "email_address")
-        else "",
-    }
+from RPA.Email.common import counter_duplicate_path
 
 
 EMAIL_CRITERIA_KEYS = {
@@ -51,8 +48,29 @@ EMAIL_CRITERIA_KEYS = {
 }
 
 
+def mailbox_to_email_address(mailbox):
+    return {
+        "name": mailbox.name if hasattr(mailbox, "name") else "",
+        "email_address": mailbox.email_address
+        if hasattr(mailbox, "email_address")
+        else "",
+    }
+
+
 class NoRecipientsError(ValueError):
     """Raised when email to be sent does not have any recipients, cc or bcc addresses."""  # noqa: E501
+
+
+class OAuth2Creds(OAuth2AuthorizationCodeCredentials):
+
+    """OAuth2 auth code flow credentials wrapper supporting token state on refresh."""
+
+    def __init__(self, *args, on_token_refresh: Callable, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._on_token_refresh = on_token_refresh
+
+    def on_token_auto_refreshed(self, access_token: OAuth2Token):
+        self._on_token_refresh(access_token)
 
 
 class Exchange:
@@ -196,24 +214,40 @@ class Exchange:
         self.account = None
         self._saved_attachments = []
 
+    def on_token_refresh(self, token: OAuth2Token):
+        """Callable you can override in order to save the newly obtained token in a
+        safe place.
+        """
+        self.logger.info(
+            "OAuth2 token was refreshed. (new expiry: %d)", token["expires_at"]
+        )
+
     def authorize(
         self,
         username: str,
-        password: str,
+        password: Optional[str] = None,
         autodiscover: Optional[bool] = True,
         access_type: Optional[str] = "DELEGATE",
         server: Optional[str] = None,
         primary_smtp_address: Optional[str] = None,
+        is_oauth: bool = False,
+        client_id: Optional[str] = None,
+        client_secret: Optional[str] = None,
+        token: Optional[dict] = None,
     ) -> None:
         """Connect to Exchange account
 
         :param username: account username
-        :param password: account password
-        :param autodiscover: use autodiscover or set it off
-        :param accesstype: default "DELEGATE", other option "IMPERSONATION"
-        :param server: required for configuration options
+        :param password: account password (can be skipped with OAuth2)
+        :param autodiscover: use autodiscover or set it off (on by default)
+        :param access_type: default "DELEGATE", other option "IMPERSONATION"
+        :param server: required for configuration setting (with `autodiscover` off)
         :param primary_smtp_address: by default set to username, but can be
-            set to be different than username
+            set to be different from username
+        :param is_oauth: use the OAuth2 authorization code flow (instead of basic auth)
+        :param client_id: registered application ID
+        :param client_secret: registered application secret (password)
+        :param token: contains access and refresh tokens, type, scope, expiry etc.
         """
         kwargs = {}
         kwargs["autodiscover"] = autodiscover
@@ -223,13 +257,37 @@ class Exchange:
         kwargs["primary_smtp_address"] = (
             primary_smtp_address if primary_smtp_address else username
         )
-        self.credentials = Credentials(username, password)
+        if is_oauth:
+            self.credentials = OAuth2Creds(
+                on_token_refresh=self.on_token_refresh,
+                identity=Identity(upn=username),
+                client_id=client_id,
+                client_secret=client_secret,
+                # Contains at least a non-expired access token or non-revoked refresh
+                #  one. (otherwise an authorization code should be present inside)
+                access_token=OAuth2Token(params=token) if token else None,
+            )
+        else:
+            if password is None:
+                raise ValueError("A password should be provided with basic auth")
+            self.credentials = Credentials(username, password)
         if server:
-            self.config = Configuration(server=server, credentials=self.credentials)
+            self.config = Configuration(
+                server=server,
+                credentials=self.credentials,
+                # Automatically detects authentication type based on the provided
+                #  `credentials` object. (when not using OAuth2)
+                auth_type=OAUTH2 if is_oauth else None,
+            )
             kwargs["config"] = self.config
         else:
             kwargs["credentials"] = self.credentials
 
+        if self.config and autodiscover:
+            self.logger.warning(
+                "Autodiscovery is left ON while using custom configuration, you may "
+                "need to turn it OFF in order to use the custom setting."
+            )
         self.account = Account(**kwargs)
 
     def list_messages(
@@ -262,6 +320,7 @@ class Exchange:
         for item in items.order_by("-datetime_received")[:count]:
             attachments = []
             if save_dir and len(item.attachments) > 0:
+                # NOTE(cmin764): Default `overwrite` param is assumed here.
                 attachments = self._save_attachments(item, save_dir)
             messages.append(self._get_email_details(item, attachments))
         return messages
@@ -288,12 +347,14 @@ class Exchange:
         messages = self.list_messages(folder_name, criterion, contains, count, save_dir)
         return [m for m in messages if not m["is_read"]]
 
-    def _get_all_items_in_folder(self, folder_name=None, parent_folder=None) -> list:
+    def _get_inbox_folder(
+        self, folder_name: str, parent_folder: Optional[str] = None
+    ) -> Inbox:
         if parent_folder is None or parent_folder is self.account.inbox:
-            target_folder = self.account.inbox / folder_name
+            parent = self.account.inbox
         else:
-            target_folder = self.account.inbox / parent_folder / folder_name
-        return target_folder.all()
+            parent = self.account.inbox / parent_folder
+        return parent / folder_name if folder_name else parent
 
     def send_message(
         self,
@@ -411,55 +472,33 @@ class Exchange:
                 if html:
                     body = body.replace(imname, f"cid:{imname}")
 
-    def create_folder(
-        self, folder_name: Optional[str] = None, parent_folder: Optional[str] = None
-    ) -> bool:
-        """Create email folder
+    def create_folder(self, folder_name: str, parent_folder: Optional[str] = None):
+        """Create email folder.
 
-        :param folder_name: name for the new folder
+        :param folder_name: name for the new folder (required)
         :param parent_folder: name for the parent folder, by default INBOX
-        :return: True if operation was successful, False if not
         """
-        if folder_name is None:
-            raise KeyError("'folder_name' is required for create folder")
-        if parent_folder is None or parent_folder is self.account.inbox:
-            parent = self.account.inbox
-        else:
-            parent = self.account.inbox / parent_folder / folder_name
-        self.logger.info(
-            "Create folder '%s'",
-            folder_name,
-        )
+        parent = self._get_inbox_folder(folder_name, parent_folder)
+        self.logger.info("Create folder %r", folder_name)
         new_folder = Folder(parent=parent, name=folder_name)
         new_folder.save()
 
-    def delete_folder(
-        self, folder_name: Optional[str] = None, parent_folder: Optional[str] = None
-    ) -> bool:
-        """Delete email folder
+    def delete_folder(self, folder_name: str, parent_folder: Optional[str] = None):
+        """Delete email folder.
 
-        :param folder_name: current folder name
+        :param folder_name: current folder name (required)
         :param parent_folder: name for the parent folder, by default INBOX
-        :return: True if operation was successful, False if not
         """
-        if folder_name is None:
-            raise KeyError("'folder_name' is required for delete folder")
-        if parent_folder is None or parent_folder is self.account.inbox:
-            folder_to_delete = self.account.inbox / folder_name
-        else:
-            folder_to_delete = self.account.inbox / parent_folder / folder_name
-        self.logger.info(
-            "Delete folder  '%s'",
-            folder_name,
-        )
+        folder_to_delete = self._get_inbox_folder(folder_name, parent_folder)
+        self.logger.info("Delete folder %r", folder_name)
         folder_to_delete.delete()
 
     def rename_folder(
         self,
-        oldname: Optional[str] = None,
-        newname: Optional[str] = None,
+        oldname: str,
+        newname: str,
         parent_folder: Optional[str] = None,
-    ) -> bool:
+    ):
         """Rename email folder
 
         :param oldname: current folder name
@@ -467,18 +506,9 @@ class Exchange:
         :param parent_folder: name for the parent folder, by default INBOX
         :return: True if operation was successful, False if not
         """
-        if oldname is None or newname is None:
-            raise KeyError("'oldname' and 'newname' are required for rename folder")
-        if parent_folder is None or parent_folder is self.account.inbox:
-            parent = self.account.inbox
-        else:
-            parent = self.account.inbox / parent_folder
-        self.logger.info(
-            "Rename folder '%s' to '%s'",
-            oldname,
-            newname,
-        )
-        items = self._get_all_items_in_folder(oldname, parent_folder)
+        parent = self._get_inbox_folder("", parent_folder)
+        self.logger.info("Rename folder %r to %r", oldname, newname)
+        items = self._get_inbox_folder(oldname, parent_folder).all()
         old_folder = Folder(parent=parent, name=oldname)
         old_folder.name = newname
         old_folder.save()
@@ -487,24 +517,22 @@ class Exchange:
 
     def empty_folder(
         self,
-        folder_name: Optional[str] = None,
+        folder_name: str,
         parent_folder: Optional[str] = None,
         delete_sub_folders: Optional[bool] = False,
-    ) -> bool:
+    ):
         """Empty email folder of all items
 
-        :param folder_name: current folder name
+        :param folder_name: current folder name (required)
         :param parent_folder: name for the parent folder, by default INBOX
         :param delete_sub_folders: delete sub folders or not, by default False
         :return: True if operation was successful, False if not
         """
-        if folder_name is None:
-            raise KeyError("'folder_name' is required for empty folder")
         if parent_folder is None:
             empty_folder = self._get_folder_object(folder_name)
         else:
             empty_folder = self._get_folder_object(f"{parent_folder} / {folder_name}")
-        self.logger.info("Empty folder '%s'", empty_folder)
+        self.logger.info("Empty folder %r", empty_folder)
         empty_folder.empty(delete_sub_folders=delete_sub_folders)
 
     def move_messages(
@@ -578,15 +606,17 @@ class Exchange:
     def _get_folder_object(self, folder_name):
         if not folder_name:
             return self.account.inbox
+
         folders = folder_name.split("/")
         if "inbox" in folders[0].lower():
             folders[0] = self.account.inbox
         folder_object = None
         for folder in folders:
             if folder_object:
-                folder_object = folder_object / folder.strip()
+                folder_object /= folder.strip()
             else:
                 folder_object = folder
+
         return folder_object
 
     def _get_filter_key_value(self, criterion):
@@ -722,6 +752,7 @@ class Exchange:
         for item in items:
             attachments = []
             if save_dir and len(item.attachments) > 0:
+                # NOTE(cmin764): Default `overwrite` param is assumed here.
                 attachments = self._save_attachments(item, save_dir)
             messages.append(self._get_email_details(item, attachments))
 
@@ -729,43 +760,61 @@ class Exchange:
             self.logger.info("Did not receive any matching items")
         return messages
 
-    def _save_attachments(self, item, save_dir, attachments_from_emls: bool = False):
+    def _save_attachments(
+        self,
+        item,
+        save_dir: str,
+        attachments_from_emls: bool = False,
+        overwrite: bool = False,
+    ):
         self._saved_attachments = self._saved_attachments or []
         incoming_items = item.attachments if hasattr(item, "attachments") else item
+
         for attachment in incoming_items:
-            self.logger.info("Attachment type: %s", type(attachment))
+            self.logger.debug("Attachment type: %s", type(attachment))
+            local_path = Path(save_dir) / attachment.name
+            if not overwrite:
+                local_path = counter_duplicate_path(local_path)
+
             if isinstance(attachment, FileAttachment):
-                local_path = os.path.join(save_dir, attachment.name)
-                with open(local_path, "wb") as f, attachment.fp as fp:
-                    buffer = fp.read(1024)
+                with open(local_path, "wb") as stream_out, attachment.fp as stream_in:
+                    buffer = stream_in.read(1024)
                     while buffer:
-                        f.write(buffer)
-                        buffer = fp.read(1024)
-                self.logger.info("Attachment saved to: %s", local_path)
-                self._saved_attachments.append(
-                    self._new_attachment_dictionary(attachment, local_path)
-                )
+                        stream_out.write(buffer)
+                        buffer = stream_in.read(1024)
             elif isinstance(attachment, ItemAttachment):
-                local_path = os.path.join(save_dir, attachment.name)
                 with open(local_path, "wb") as message_out:
                     message_out.write(attachment.item.mime_content)
-                self.logger.info("Attachment saved to: %s", local_path)
-                self._saved_attachments.append(
-                    self._new_attachment_dictionary(attachment, local_path)
-                )
                 if attachments_from_emls:
-                    self._save_attachments(attachment.item, save_dir, False)
+                    self._save_attachments(
+                        attachment.item,
+                        save_dir,
+                        attachments_from_emls=False,
+                        overwrite=overwrite,
+                    )
+            else:
+                self.logger.warning(
+                    "Unrecognized attachment type: %s", type(attachment)
+                )
+                return self._saved_attachments
+
+            self.logger.info("Attachment saved to: %s", local_path)
+            self._saved_attachments.append(
+                self._new_attachment_dictionary(attachment, local_path)
+            )
+
         return self._saved_attachments
 
-    def _new_attachment_dictionary(self, attachment, local_path):
+    @staticmethod
+    def _new_attachment_dictionary(
+        attachment: Union[FileAttachment, ItemAttachment], local_path: Path
+    ) -> Dict:
         return {
             "name": attachment.name,
             "content_type": attachment.content_type,
             "size": attachment.size,
-            "is_contact_photo": attachment.is_contact_photo
-            if hasattr(attachment, "is_contact_photo")
-            else False,
-            "local_path": local_path,
+            "is_contact_photo": getattr(attachment, "is_contact_photo", False),
+            "local_path": str(local_path),
         }
 
     def _get_email_details(self, item, attachments):
@@ -802,39 +851,52 @@ class Exchange:
         self,
         message: Union[dict, str],
         save_dir: Optional[str] = None,
-        attachments_from_emls: Optional[bool] = False,
+        attachments_from_emls: bool = False,
+        overwrite: bool = False,
     ) -> list:
-        """Save attachments in message into given directory
+        """Save attachments from message into given directory.
 
-        :param message: dictionary or .eml filepath containing message details
-        :param save_dir: filepath where attachments will be saved
-        :param attachments_from_emls: if attachment is a EML file, set to True to
-         save attachments from that EML file, default False
+        :param message: dictionary or .eml file path containing message details
+        :param save_dir: file path where attachments will be saved
+        :param attachments_from_emls: pass `True` if the attachment is an EML file (for
+            saving attachments from that EML file instead), `False` otherwise (default)
+        :param overwrite: overwrite existing downloaded attachments with the same name
+            if set to `True`, `False` otherwise (default)
         :return: list of saved attachments
 
-        Example.
+        Example:
 
         .. code:: robotframework
 
-            ${messages}=    List Messages
-            FOR    ${msg}    IN    @{messages}}
-                Save Attachments    ${msg}    %{ROBOT_ARTIFACTS}  True
+            ${messages} =    List Messages
+            FOR    ${msg}    IN    @{messages}
+                Save Attachments    ${msg}    %{ROBOT_ARTIFACTS}
+                ...    attachments_from_emls=${True}
             END
-            ${attachments}=  Save Attachments  ${CURDIR}${/}saved.eml  %{ROBOT_ARTIFACTS}
+
+            ${attachments} =    Save Attachments    ${CURDIR}${/}saved.eml
+            ...    %{ROBOT_ARTIFACTS}    overwrite=${True}
         """  # noqa: E501
         self._saved_attachments = []
         if isinstance(message, dict):
             return self._save_attachments(
-                message["attachments_object"], save_dir, attachments_from_emls
+                message["attachments_object"],
+                save_dir,
+                attachments_from_emls=attachments_from_emls,
+                overwrite=overwrite,
             )
         else:
             # extract attachments from .eml file
             absolute_filepath = Path(message).resolve()
             if absolute_filepath.suffix != ".eml":
                 raise ValueError("Filename extension needs to be '.eml'")
-            return self._save_attachments_from_file(message, save_dir)
+            return self._save_attachments_from_file(
+                message, save_dir, overwrite=overwrite
+            )
 
-    def _save_attachments_from_file(self, filename: str, save_dir: str):
+    def _save_attachments_from_file(
+        self, filename: str, save_dir: str, *, overwrite: bool
+    ):
         """
         Try to extract the attachments from given .eml file
         """
@@ -848,7 +910,9 @@ class Exchange:
                 output_filename = attachment.get_filename()
                 # If no attachments are found, skip this file
                 if output_filename:
-                    local_path = os.path.join(save_dir, output_filename)
+                    local_path = Path(save_dir) / output_filename
+                    if not overwrite:
+                        local_path = counter_duplicate_path(local_path)
                     with open(local_path, "wb") as of:
                         payload = attachment.get_payload(decode=True)
                         of.write(payload)
@@ -858,15 +922,15 @@ class Exchange:
                                 "content_type": None,
                                 "size": len(payload),
                                 "is_contact_photo": None,
-                                "local_path": local_path,
+                                "local_path": str(local_path),
                             }
                         )
             if len(attachments) == 0:
                 self.logger.warning("No attachment found for file %s!", f.name)
         return attachments
 
-    def save_message(self, message: dict, filename: str) -> list:
-        """Save email as .eml file
+    def save_message(self, message: dict, filename: str):
+        """Save email as .eml file.
 
         :param message: dictionary containing message details
         :param filename: name of the file to save message into